requirements.txt
Flask==2.3.2
pyTelegramBotAPI==4.12.0